# dreamComeTrueBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dream Come True Information Technology Api Docs",
        "version": "v1"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "products api",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "count of rows in a single page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "test product",
                        "required": false,
                        "schema": {
                            "type": "text"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Invalid Request Validation"
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Store product",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_name",
                                    "price",
                                    "category",
                                    "stock_status"
                                ],
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Test Product"
                                    },
                                    "price": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "230"
                                    },
                                    "category": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "cars"
                                    },
                                    "stock_status": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "422": {
                        "description": "Invalid Request Validation"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "products",
            "description": "products"
        }
    ]
}